LayoutTests/imported/w3c/web-platform-tests/css/selectors/focus-visible-006-expected.txt

This test checks that :focus-visible always matches on elements with contenteditable=true set.
If the user-agent does not claim to support the :focus-visible pseudo-class then SKIP this test.
Click the content editable span below to focus it.
If the element has a red background, then the test result is FAILURE. If the element has a green outline, then the test result is SUCCESS.

Focus me

FAIL Focus should always match :focus-visible on content editable divs assert_equals: outlineColor for SPAN#el should be green expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
